Fix CountFalsePositives so it a) does not return garbage for target territory b) returns a proper fraction for false positive rate (#3228)

+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/validation/CountFalsePositives.java
@@ -75,7 +75,7 @@ public class CountFalsePositives extends VariantWalker {
     @Override
     public Object onTraversalSuccess() {
         final List<SimpleInterval> intervals =  intervalArgumentCollection.getIntervals(getReferenceDictionary());
-        final long targetTerritory = intervals.stream().mapToInt(i -> i.size()).sum();
+        final long targetTerritory = intervals.stream().mapToLong(i -> i.size()).sum();
 
         try ( FalsePositiveTableWriter writer = new FalsePositiveTableWriter(outputFile) ) {
             FalsePositiveRecord falsePositiveRecord = new FalsePositiveRecord(id, snpFalsePositiveCount, indelFalsePositiveCount, targetTerritory);
